+++ src/app/lib/views/browser/favorite_browser.js
@@ -2,7 +2,11 @@
     'use strict';
 
     var FavoriteBrowser = App.View.PCTBrowser.extend({
-        collectionModel: App.Model.FavoriteCollection
+        collectionModel: App.Model.FavoriteCollection,
+        filters: {
+            genres: App.Config.genres,
+            sorters: App.Config.sorters
+        }
     });
 
     App.View.FavoriteBrowser = FavoriteBrowser;